Protecting against timing attacks
The previous string comparison method didn't operate in constant time.
If the string under test wasn't the same length as the secret the method
instantly returned.

This implementation is constant time with respect to the untrusted
string and executes very quickly (that's why the style is not quite as compact
as one might expect).